% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iet.R
\name{classify_tsi_carlson}
\alias{classify_tsi_carlson}
\title{Classifica TSI (Carlson) em faixas qualitativas}
\usage{
classify_tsi_carlson(x, locale = c("pt", "en"))
}
\arguments{
\item{x}{Vetor numerico com TSI/IET (0-100). NA preservado.}

\item{locale}{Idioma dos rotulos: \code{"pt"} (padrao) ou \code{"en"}.}
}
\value{
Fator ordenado com classes de trofia.
}
\description{
Converte valores do indice trofico de Carlson (TSI/IET) para classes
qualitativas ordenadas. Retorna fator ordenado em portugues ("pt")
ou ingles ("en").
}
\examples{
classify_tsi_carlson(c(25, 35, 45, 60, 80))
classify_tsi_carlson(c(25, 35, 45, 60, 80), locale = "en")

}
