% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iet_lamparelli.R
\name{classify_tsi_lamparelli}
\alias{classify_tsi_lamparelli}
\title{Classifica TSI (Lamparelli) em faixas qualitativas}
\usage{
classify_tsi_lamparelli(x, locale = c("pt", "en"))
}
\arguments{
\item{x}{Vetor numerico com TSI/IET (0-100). NA preservado.}

\item{locale}{Idioma dos rotulos: \code{"pt"} (padrao) ou \code{"en"}.}
}
\value{
Fator ordenado com classes de trofia (Lamparelli).
}
\description{
Converte valores do indice trofico de Lamparelli (TSI/IET) para classes
qualitativas ordenadas. Retorna fator ordenado em portugues ("pt")
ou ingles ("en").
}
\examples{
classify_tsi_lamparelli(c(40, 50, 56, 61, 65, 72))
classify_tsi_lamparelli(c(40, 50, 56, 61, 65, 72), locale = "en")

}
