% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conama.R
\name{conama_summary}
\alias{conama_summary}
\title{Conformity summary (long format)}
\usage{
conama_summary(df, classe = "2")
}
\arguments{
\item{df}{dados de entrada}

\item{classe}{classe CONAMA}
}
\value{
tibble com: parametro, valor, lim_min, lim_max, status, ok, delta
}
\description{
Conformity summary (long format)
}
