% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conama.R
\name{conama_text}
\alias{conama_text}
\title{Text summary of conformity (bulleted, formatted)}
\usage{
conama_text(
  df,
  classe = "2",
  only_violations = FALSE,
  decimal_mark = ",",
  big_mark = "."
)
}
\arguments{
\item{df}{dados de entrada}

\item{classe}{classe CONAMA}

\item{only_violations}{se TRUE, lista apenas parametros com violacao}

\item{decimal_mark}{separador decimal (default ",")}

\item{big_mark}{separador de milhar (default ".")}
}
\value{
character vector (linhas)
}
\description{
Text summary of conformity (bulleted, formatted)
}
