% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_text.R
\name{generate_analysis}
\alias{generate_analysis}
\title{Geração de parágrafos analíticos (sem IA)}
\usage{
generate_analysis(
  df,
  classe_conama = "2",
  incluir_tendencia = TRUE,
  parametros_tendencia = c("turbidez", "od", "pH"),
  contexto = list(rio = NA, periodo = NA, cidade = NA)
)
}
\description{
Geração de parágrafos analíticos (sem IA)
}
