% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iet.R
\name{iet_carlson}
\alias{iet_carlson}
\title{Índice do Estado Trófico (Carlson) - básico}
\usage{
iet_carlson(secchi = NULL, clorofila = NULL, tp = NULL)
}
\description{
Índice do Estado Trófico (Carlson) - básico
}
