% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iet_lamparelli.R
\name{iet_lamparelli}
\alias{iet_lamparelli}
\title{IET - Lamparelli (estrutura v0.2.1)}
\usage{
iet_lamparelli(
  tp = NULL,
  chla = NULL,
  sd = NULL,
  ambiente = c("rio", "reservatorio")
)
}
\description{
IET - Lamparelli (estrutura v0.2.1)
}
