% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap of parameters vs. sites}
\usage{
plot_heatmap(df_long)
}
\arguments{
\item{df_long}{Long-format data frame (site/parameter/value).}
}
\value{
A ggplot object (class \code{ggplot}) representing the heatmap.
}
\description{
Heatmap of parameters vs. sites
}
