% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqa_curves.R
\name{qi_interp}
\alias{qi_interp}
\title{Interpola Qi a partir de uma curva definida por pontos (x, qi)}
\usage{
qi_interp(x, tbl)
}
\description{
Interpola Qi a partir de uma curva definida por pontos (x, qi)
}
\keyword{internal}
