% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_clean.R
\name{validate_wq}
\alias{validate_wq}
\title{Valida presenca de colunas minimas}
\usage{
validate_wq(
  df,
  required = c("ph", "turbidez", "od", "dbo", "nt_total", "p_total", "tds",
    "temperatura", "coliformes")
)
}
\arguments{
\item{df}{data.frame/tibble de entrada.}

\item{required}{vetor de nomes de colunas obrigatorias.}
}
\value{
O \code{df} de entrada se estiver valido; caso contrario, erro.
}
\description{
Garante que o conjunto minimo de colunas exista no dataset.
}
