% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqa.R
\name{classify_iqa}
\alias{classify_iqa}
\title{Classifica valores do IQA/WQI em faixas qualitativas}
\usage{
classify_iqa(x, locale = c("pt", "en"))
}
\arguments{
\item{x}{Vetor numerico com IQA em 0-100. Valores NA sao preservados.}

\item{locale}{Idioma dos rotulos: \code{"pt"} (padrao) ou \code{"en"}.}
}
\value{
Um fator ordenado com os rotulos de classe.
}
\description{
Converte valores numericos de IQA (0-100) em classes qualitativas
padronizadas. Suporta rotulos em portugues ("pt") ou ingles ("en").
}
\examples{
classify_iqa(c(15, 40, 65, 80, 95))
classify_iqa(c(15, 40, 65, 80, 95), locale = "en")

}
