% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iet_lamparelli.R
\name{iet_lamparelli}
\alias{iet_lamparelli}
\title{Trophic State Index - Lamparelli (skeleton v0.2.1)}
\usage{
iet_lamparelli(
  tp = NULL,
  chla = NULL,
  sd = NULL,
  ambiente = c("rio", "reservatorio")
)
}
\arguments{
\item{tp}{Numeric total phosphorus (mg/L)}

\item{chla}{Numeric chlorophyll-a (µg/L)}

\item{sd}{Numeric Secchi disk depth (m)}

\item{ambiente}{Environment type: 'rio' or 'reservatorio'}
}
\value{
Data frame with IET components and overall Lamparelli index
}
\description{
Trophic State Index - Lamparelli (skeleton v0.2.1)
}
