% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsfwqi.R
\name{nsfwqi}
\alias{nsfwqi}
\title{NSF Water Quality Index (prototype v0.2.1)}
\usage{
nsfwqi(
  df,
  pesos = c(do = 0.17, fc = 0.16, ph = 0.11, bod = 0.11, temp_change = 0.1, po4 = 0.1,
    no3 = 0.1, turbidez = 0.08, sst = 0.07),
  na_rm = FALSE
)
}
\arguments{
\item{df}{Data frame containing columns compatible with NSFWQI mapping}

\item{pesos}{Named numeric vector with parameter weights}

\item{na_rm}{Logical; allow NA and rescale weights}
}
\value{
Data frame with an added 'NSFWQI' column
}
\description{
NSF Water Quality Index (prototype v0.2.1)
}
