% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_box}
\alias{plot_box}
\title{Boxplots by site/parameter}
\usage{
plot_box(df, parametro, by = "ponto")
}
\arguments{
\item{df}{Data frame with water quality data.}

\item{parametro}{Character; name of the numeric parameter column.}

\item{by}{Character; grouping column (e.g., "ponto").}
}
\value{
A \code{ggplot} object.
}
\description{
Boxplots of one numeric parameter grouped by a categorical column.
}
\examples{
\donttest{
data(wq_demo)
plot_box(wq_demo, "turbidez", by = "ponto")
}

}
\seealso{
\code{\link[=plot_series]{plot_series()}}, \code{\link[=plot_heatmap]{plot_heatmap()}}, \code{\link[=iqa]{iqa()}}
}
