% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_iqa}
\alias{plot_iqa}
\title{Plot IQA by site/date}
\usage{
plot_iqa(df)
}
\arguments{
\item{df}{Data frame returned by \code{iqa()} (or with equivalent columns).}
}
\value{
A \code{ggplot} object.
}
\description{
Bar plot of IQA values per site/date. Requires an \code{IQA} column.
}
\examples{
\donttest{
data(wq_demo)
d <- iqa(wq_demo, na_rm = TRUE)
plot_iqa(d)
}

}
\seealso{
\code{\link[=iqa]{iqa()}}, \code{\link[=plot_series]{plot_series()}}, \code{\link[=plot_box]{plot_box()}}
}
