% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_report}
\alias{render_report}
\title{Gera relatório (R Markdown / Quarto) a partir de template}
\usage{
render_report(
  df,
  meta = list(rio = "Buranhém", periodo = "2025-01"),
  output = "wq_report.html",
  template = system.file("templates", "report_qmd.qmd", package = "tikatuwq")
)
}
\description{
Gera relatório (R Markdown / Quarto) a partir de template
}
