% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikatuwq-package.R
\docType{package}
\name{tikatuwq-package}
\alias{tikatuwq}
\alias{tikatuwq-package}
\title{tikatuwq: Ferramentas para analise de qualidade da agua (Brasil)}
\description{
Indices (IQA CETESB/NSF, IET Carlson/Lamparelli), checagem CONAMA 357/2005,
visualizacoes e texto analitico automatico (rule-based).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tikatuwq/tikatuwq}
  \item \url{https://tikatuwq.github.io/tikatuwq/}
  \item Report bugs at \url{https://github.com/tikatuwq/tikatuwq/issues}
}

}
\author{
\strong{Maintainer}: Vinicius Saraiva Santos \email{vinisaraiva@gmail.com} (\href{https://orcid.org/0009-0007-1387-7927}{ORCID})

}
\keyword{internal}
